refactor: 파일 업로드 시 디렉토리 수정

+++ src/main/java/com/wootube/ioi/service/VideoService.java
@@ -1,66 +1,67 @@
 package com.wootube.ioi.service;
 
+import java.util.List;
+import javax.transaction.Transactional;
+
 import com.wootube.ioi.domain.model.User;
 import com.wootube.ioi.domain.model.Video;
-import com.wootube.ioi.domain.repository.UserRepository;
 import com.wootube.ioi.domain.repository.VideoRepository;
 import com.wootube.ioi.service.dto.VideoRequestDto;
 import com.wootube.ioi.service.dto.VideoResponseDto;
-import com.wootube.ioi.service.exception.NotFoundUserException;
-import com.wootube.ioi.service.exception.NotFoundVideoException;
+import com.wootube.ioi.service.exception.NotFoundVideoIdException;
+import com.wootube.ioi.service.exception.NotMatchUserIdException;
+import com.wootube.ioi.service.exception.UserAndWriterMisMatchException;
 import com.wootube.ioi.service.util.FileUploader;
+import com.wootube.ioi.service.util.UploadType;
 import org.modelmapper.ModelMapper;
+
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.transaction.Transactional;
-import java.util.List;
-
 @Service
 public class VideoService {
+
     private final FileUploader fileUploader;
     private final ModelMapper modelMapper;
 
     private final VideoRepository videoRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
-    public VideoService(FileUploader fileUploader, ModelMapper modelMapper, VideoRepository videoRepository, UserRepository userRepository) {
+    public VideoService(FileUploader fileUploader, ModelMapper modelMapper, VideoRepository videoRepository, UserService userService) {
         this.fileUploader = fileUploader;
         this.modelMapper = modelMapper;
         this.videoRepository = videoRepository;
-        this.userRepository = userRepository;
+        this.userService = userService;
     }
 
-    public VideoResponseDto create(MultipartFile uploadFile, VideoRequestDto videoRequestDto, Long userId) {
-        String videoUrl = fileUploader.uploadFile(uploadFile);
+    public VideoResponseDto create(MultipartFile uploadFile, VideoRequestDto videoRequestDto) {
+        String videoUrl = fileUploader.uploadFile(uploadFile, UploadType.VIDEO);
         String originFileName = uploadFile.getOriginalFilename();
 
         Video video = modelMapper.map(videoRequestDto, Video.class);
-        User writer = findUser(userId);
+        User writer = userService.findByIdAndIsActiveTrue(videoRequestDto.getWriterId());
 
         video.initialize(videoUrl, originFileName, writer);
         return modelMapper.map(videoRepository.save(video), VideoResponseDto.class);
     }
 
-    private User findUser(Long userId) {
-        return userRepository.findById(userId).orElseThrow(NotFoundUserException::new);
+    public VideoResponseDto findVideo(Long id) {
+        return modelMapper.map(findById(id), VideoResponseDto.class);
     }
 
-    public VideoResponseDto findById(Long id) {
-        return modelMapper.map(findVideo(id), VideoResponseDto.class);
-    }
-
-    public Video findVideo(Long id) {
+    public Video findById(Long id) {
         return videoRepository.findById(id)
-                .orElseThrow(NotFoundVideoException::new);
+                .orElseThrow(NotFoundVideoIdException::new);
     }
 
     @Transactional
     public void update(Long id, MultipartFile uploadFile, VideoRequestDto videoRequestDto) {
-        Video video = findVideo(id);
+        Video video = findById(id);
+        matchWriter(videoRequestDto.getWriterId(), id);
+
         if (!uploadFile.isEmpty()) {
-            fileUploader.deleteFile(video.getOriginFileName());
-            String videoUrl = fileUploader.uploadFile(uploadFile);
+            fileUploader.deleteFile(video.getOriginFileName(), UploadType.VIDEO);
+            String videoUrl = fileUploader.uploadFile(uploadFile, UploadType.VIDEO);
             video.updateContentPath(videoUrl);
         }
         video.update(modelMapper.map(videoRequestDto, Video.class));
@@ -68,16 +69,24 @@ public class VideoService {
     }
 
     @Transactional
-    public void deleteById(Long videoId) {
-        Video video = findVideo(videoId);
-//        Video writerVideo = video.matchWriter(userId);
-
-        fileUploader.deleteFile(video.getOriginFileName());
+    public void deleteById(Long videoId, Long userId) {
+        Video video = findById(videoId);
+        if (!video.matchWriter(userId)) {
+            throw new UserAndWriterMisMatchException();
+        }
+        fileUploader.deleteFile(video.getOriginFileName(), UploadType.VIDEO);
         videoRepository.deleteById(video.getId());
     }
 
-    public List<Video> findVideosByWriter(Long writerId) {
-        User writer = findUser(writerId);
+    public List<Video> findAllByWriter(Long writerId) {
+        User writer = userService.findByIdAndIsActiveTrue(writerId);
         return videoRepository.findByWriter(writer);
+    }
+
+    public void matchWriter(Long userId, Long videoId) {
+        Video video = findById(videoId);
+        if (!video.matchWriter(userId)) {
+            throw new NotMatchUserIdException();
+        }
     }
 }
